Events.py Creation
Made a new file specifically for event functionality. This further simplifies dashboard.py and init.py
